requirements.txt
pytest==3.0.6
docopt==0.6.2
ExifRead==2.1.2
httpie==0.9.9
path.py==10.1
-e git+https://github.com/OpenPathView/DirectoryManagerClient@stable#egg=opv_directorymanagerclient
-e git+https://github.com/OpenPathView/OPV_DBRest-client@stable#egg=opv_api_client
pyudev==0.21.0
geojson==1.3.4
PyYAML==3.12

